Website created to apply knowledge learned during the fnb app academy bootcamp.
So the fnb app academy is a programming bootcamp - it's 9 weeks long and offers various tutorials and lessons with the aim of seeding more South African app-preneurs into the tech-space. I applied and this is the website I decided to create to help me apply everything I learn from the bootcamp as well as on my own.